% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.r
\name{build_tree}
\alias{build_tree}
\title{title "build_tree"
description "build_tree"}
\usage{
build_tree(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
A tree structure
}
\description{
title "build_tree"
description "build_tree"
}
