% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.r
\name{calc_leaf}
\alias{calc_leaf}
\title{Calculate the best split leaf value for a given data frame and target variable}
\usage{
calc_leaf(data, target, type)
}
\arguments{
\item{data}{A data frame}

\item{target}{The target variable}

\item{type}{The type of the target variable, either "classification" or "regression"}
}
\value{
A list containing the best split leaf value and the corresponding Gini impurity
}
\description{
Calculate the best split leaf value for a given data frame and target variable
}
