% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.r
\name{calculate_accuracy}
\alias{calculate_accuracy}
\title{Calculate the accuracy of a random forest}
\usage{
calculate_accuracy(forest, X_test, y_test)
}
\arguments{
\item{forest}{A random forest object.}

\item{X_test}{A data frame containing the test data.}

\item{y_test}{A vector containing the true labels for the test data.}
}
\value{
The accuracy of the random forest.
}
\description{
Calculate the accuracy of a random forest
}
