% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.r
\name{calculate_gini}
\alias{calculate_gini}
\title{Calculate the Gini impurity of a vector}
\usage{
calculate_gini(y)
}
\arguments{
\item{y}{A vector}
}
\value{
The Gini impurity of the vector
}
\description{
Calculate the Gini impurity of a vector
}
