% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.r
\name{collapse}
\alias{collapse}
\title{title "collapse"
description "collapse"}
\usage{
collapse(x)
}
\arguments{
\item{x}{a vactor}
}
\value{
x with spaces replaced by underscores
}
\description{
title "collapse"
description "collapse"
}
