% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{ensemble_predict}
\alias{ensemble_predict}
\title{Imitate randomforest}
\usage{
ensemble_predict(
  train_data,
  test_data,
  target_variable,
  num_trees = 100,
  label_mapping = NULL
)
}
\arguments{
\item{train_data}{we split the original data randomly into 2 datasets, one is train data,usually 70\% of the data}

\item{test_data}{we split the original data randomly into 2 datasets, another is test data,usually 30\% of the data}

\item{target_variable}{the dependent variable that we would like to predict}

\item{num_trees}{number of the decision trees we create}

\item{label_mapping}{whether there are special algorithm for the target variable,eg for penguins:"1" = "Adelie", "2" = "Chinstrap", "3" = "Gentoo"}
}
\value{
prediction of the target variable based on the training dataset
}
\description{
Imitate randomforest
}
\examples{
library(palmerpenguins)
library(tidyverse)
library(dplyr)
library(stringr)
library(randomForest)
library(rpart)
data("penguins")
penguins <- na.omit(penguins)
penguins <- penguins \%>\% select(-year)
set.seed(123)
sample_indices <- sample(1:nrow(penguins), 0.7 * nrow(penguins))
train_data <- penguins[sample_indices, ]
test_data <- penguins[-sample_indices, ]
target_variable <- "species"
ensemble_pred <- ensemble_predict(train_data, test_data, target_variable)
}
