% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.r
\name{find_best_split_classification}
\alias{find_best_split_classification}
\title{Find the best split for classification}
\usage{
find_best_split_classification(data, features, target, min_samples_leaf)
}
\arguments{
\item{data}{A data frame}

\item{features}{A vector of feature names}

\item{target}{The target variable name}

\item{min_samples_leaf}{The minimum number of samples required to be at a leaf node}
}
\value{
A list containing the best split feature, value, and gini impurity
}
\description{
Find the best split for classification
}
