% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.r
\name{find_best_split_extratrees}
\alias{find_best_split_extratrees}
\title{title "Find the best split for extratrees"
description "Find the best split for extratrees"}
\usage{
find_best_split_extratrees(data, features, target)
}
\arguments{
\item{data}{A data frame}

\item{features}{A vector of feature names}

\item{target}{The target variable name}
}
\value{
A list containing the best split feature, value, and gini impurity
}
\description{
title "Find the best split for extratrees"
description "Find the best split for extratrees"
}
