% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.r
\name{plot_tree}
\alias{plot_tree}
\title{plot_tree}
\usage{
plot_tree(tree, tree_from = "ranger")
}
\arguments{
\item{tree}{the tree info dataframe}

\item{tree_from}{the method of making tree}
}
\value{
A tree structure
}
\description{
plot_tree
}
