% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{rf_cv_function}
\alias{rf_cv_function}
\title{use cross-validation to tune hyperparameter(complexity parameter)}
\usage{
rf_cv_function(train_data, target)
}
\arguments{
\item{train_data}{we split the original data randomly into 2 datasets, one is train data,usually 70\% of the data}

\item{target}{the dependent variable that we would like to predict}
}
\value{
contains information about the cross-validated tuning process and the best-tuned random forest model
}
\description{
use cross-validation to tune hyperparameter(complexity parameter)
}
\examples{
library(palmerpenguins)
library(tidyverse)
library(dplyr)
library(stringr)
library(randomForest)
library(rpart)
library(caret)
data("penguins")
penguins <- na.omit(penguins)
penguins <- penguins \%>\% select(-year)
set.seed(123)
sample_indices <- sample(1:nrow(penguins), 0.7 * nrow(penguins))
train_data <- penguins[sample_indices, ]
rf_cv <- rf_cv_function(train_data, 'species')
}
