% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.r
\name{tree_info}
\alias{tree_info}
\title{get the information of a tree}
\usage{
tree_info(forest, tree_number = 1)
}
\arguments{
\item{forest}{A list of trees}

\item{tree_number}{The number of the tree to get information from}
}
\value{
A data frame containing the information of the tree
}
\description{
get the information of a tree
}
