% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.r
\name{build_decision_tree}
\alias{build_decision_tree}
\title{Build Decision Tree}
\usage{
build_decision_tree(
  X,
  y,
  max_depth = Inf,
  min_samples_split = 2,
  min_samples_leaf = 1,
  type = "classification"
)
}
\arguments{
\item{X}{a data frame of features}

\item{y}{a vector of target values}

\item{max_depth}{the maximum depth of the tree}

\item{min_samples_split}{the minimum number of samples required to split an internal node}

\item{min_samples_leaf}{the minimum number of samples required to be at a leaf node}

\item{type}{the type of the tree, either "classification" or "regression"}
}
\value{
a decision tree
}
\description{
Build a decision tree for classification or regression
}
\examples{
# Build a classification tree
X <- data.frame(x1 = c(1, 2, 3, 4, 5), x2 = c(1, 2, 3, 4, 5))
y <- c(1, 0, 1, 0, 1)
tree <- build_decision_tree(X, y, type = "classification")
# Build a regression tree
X <- data.frame(x1 = c(1, 2, 3, 4, 5), x2 = c(1, 2, 3, 4, 5))
y <- c(1, 2, 3, 4, 5)
tree <- build_decision_tree(X, y, type = "regression")
}
