% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.r
\name{predict_random_forest}
\alias{predict_random_forest}
\title{title "Predict using a trained random forest"}
\usage{
predict_random_forest(forest, new_data, n_cores = availableCores() - 1)
}
\arguments{
\item{forest}{A random forest object.}

\item{new_data}{A data frame containing the new data to be predicted.}

\item{n_cores}{The number of cores to use for parallel processing. Default is the number of available cores minus one.}
}
\value{
A vector of predictions.
}
\description{
title "Predict using a trained random forest"
}
