% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.r
\name{predict_tree}
\alias{predict_tree}
\title{Predict using decision tree}
\usage{
predict_tree(tree, new_data)
}
\arguments{
\item{tree}{decision tree}

\item{new_data}{new data}
}
\value{
the predict result
}
\description{
Predict using decision tree
}
